requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pandas
openpyxl
jinja2
gunicorn
